glog/requirements.txt
agora-api==0.26.0
agora-token-builder==1.0.0
aiohttp==3.8.3
aiosignal==1.3.1
amqp==5.1.1
asgiref==3.5.2
astroid==2.13.3
async-generator==1.10
async-timeout==4.0.2
attrs==21.3.0
autobahn==22.2.2
Automat==20.2.0
autopep8==2.0.1
Babel==2.11.0
backoff==1.11.1
beautifulsoup4==4.10.0
billiard==3.6.4.0
bitcoin==1.1.42
bleach==6.0.0
branca==0.4.2
Brotli==1.0.9
caer==2.0.8
celery==5.2.7
certifi==2021.10.8
cffi==1.15.0
channels==3.0.5
channels-redis==4.0.0
chardet==5.1.0
charset-normalizer==2.0.9
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
constantly==15.1.0
coverage==6.4.2
crispy-bootstrap5==0.7
cryptography==36.0.1
cssselect==1.2.0
cssselect2==0.7.0
cycler==0.11.0
daphne==3.0.2
DateTime==4.3
decorator==5.1.0
defusedxml==0.7.1
dicttoxml==1.7.16
dill==0.3.6
distlib==0.3.4
Django==4.1
django-active-link==0.1.8
django-admin-confirm==0.2.5
django-admin-material-dashboard==1.0.8
django-agora==0.4
django-allauth==0.52.0
django-appconf==1.0.5
django-audio-validator==0.0.3
django-braces==1.15.0
django-celery-beat==2.1.0
django-cities==0.6.1
django-cors-headers==3.10.1
django-countries==7.5
django-crispy-forms==1.14.0
django-embed-video==1.4.8
django-environ==0.9.0
django-filter==21.1
django-jazzmin==2.6.0
django-localflavor==3.1
django-material-admin==1.8.6
django-model-utils==4.3.1
django-mssql-backend==2.8.1
django-parler==2.3
django-phone-field==1.8.1
django-phonenumber-field==7.0.1
django-redis==5.2.0
django-redisboard==8.4.0
django-referral==1.0.1
django-reflinks==0.5.1
django-rosetta==0.9.8
django-sendgrid-v5==1.2.1
django-ssl-smtp==2.0.0
django-taggit==3.1.0
django-timezone-field==4.2.3
django-widget-tweaks==1.4.12
django_debug_toolbar==3.8.1
djangorestframework==3.13.1
docutils==0.19
easy-thumbnails==2.8.5
environ==1.0
filelock==3.4.2
filetype==1.2.0
flower==1.2.0
folium==0.12.1.post1
fonttools==4.28.5
frozenlist==1.3.3
future==0.18.2
geocoder==1.38.1
geoip2==4.6.0
geolocation-python==0.2.2
h11==0.12.0
html5lib==1.1
humanize==4.4.0
hyperlink==21.0.0
idna==3.3
importlib-metadata==6.0.0
incremental==21.3.0
ip2geotools==0.1.6
IP2Location==8.9.0
isort==5.11.4
jaraco.classes==3.2.3
Jinja2==3.0.3
keyring==23.13.1
kiwisolver==1.3.2
kombu==5.2.4
lazy-object-proxy==1.9.0
lxml==4.7.1
Markdown==3.4.1
markdown-it-py==2.1.0
MarkupSafe==2.0.1
matplotlib==3.5.1
maxminddb==2.2.0
mccabe==0.7.0
mdurl==0.1.2
more-itertools==9.0.0
MouseInfo==0.1.3
msgpack==1.0.4
multidict==6.0.4
mypy==0.930
mypy-extensions==0.4.3
mysql-connector-python==8.0.30
mysqlclient==2.1.1
nose==1.3.7
numpy==1.22.0rc3
oauthlib==3.2.2
opencage==2.0.0
opencv-contrib-python==4.5.5.62
opencv-python==4.5.4.60
outcome==1.1.0
packaging==21.3
pandas==1.3.5
phone-iso3166==0.3.8
phonenumbers==8.12.40
Pillow==9.4.0
pinax-referrals==4.0.1
pip-review==1.3.0
pkginfo==1.9.6
platformdirs==2.4.1
playsound==1.3.0
playsound2==0.1
polib==1.1.1
prometheus-client==0.15.0
prompt-toolkit==3.0.33
protobuf==3.19.4
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.53
pycodestyle==2.10.0
pycountry==20.7.3
pycountrycode==1.1.7
pycparser==2.21
pydyf==0.5.0
pygame==2.1.2
PyGetWindow==0.0.9
Pygments==2.14.0
PyJWT==2.6.0
pylint==2.15.10
pymemcache==4.0.0
PyMsgBox==1.0.9
PyMySQL==1.0.2
pyodbc==4.0.34
pyOpenSSL==21.0.0
pyparsing==3.0.6
pyperclip==1.8.2
pyphen==0.13.2
pypng==0.0.21
PyQRCode==1.2.1
pyquery==2.0.0
PyRect==0.1.4
PyScreeze==0.1.28
python-barcode==0.13.1
python-crontab==2.6.0
python-dateutil==2.8.1
python-decouple==3.6
python-http-client==3.3.7
python-memcached==1.59
python-stdnum==1.18
python3-openid==3.2.0
pytweening==1.0.4
pytz==2021.3
pywin32-ctypes==0.2.0
qrcode==7.3.1
rabbitmq==0.2.0
ratelim==0.1.6
readme-renderer==37.3
redis==4.4.2
requests==2.26.0
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
rfc3986==2.0.0
rich==13.2.0
schedule==1.1.0
scipy==1.8.0rc2
selenium==4.1.0
sendgrid==6.9.7
service-identity==21.1.0
sib-api-v3-sdk==7.5.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
sounddevice==0.4.3
soupsieve==2.3.1
sqlparse==0.4.2
starkbank-ecdsa==2.0.3
stripe==5.0.0
swapper==1.3.0
tinycss2==1.2.1
tk==0.1.0
tomli==2.0.0
tomlkit==0.11.6
tornado==6.2
tqdm==4.64.1
trio==0.19.0
trio-websocket==0.9.2
twine==4.0.2
Twisted==22.2.0
twisted-iocpsupport==1.0.2
txaio==22.2.1
typed-ast==1.5.4
typing==3.7.4.3
typing_extensions==4.0.1
tzdata==2021.5
urllib3==1.26.7
vine==5.0.0
virtualenv==20.11.0
wcwidth==0.2.5
weasyprint==57.2
webencodings==0.5.1
wrapt==1.14.1
wsproto==1.0.0
xlrd==2.0.1
xmltodict==0.12.0
yarl==1.8.2
zipp==3.11.0
zope.interface==5.4.0
zopfli==0.2.2
